Separated llms from the main llm.py file and fixed some deprication bugs
