+++ examples/lightrag_api_open_webui_demo.py
@@ -9,7 +9,7 @@ from typing import Optional
 import os
 import logging
 from lightrag import LightRAG, QueryParam
-from lightrag.llm import ollama_model_complete, ollama_embed
+from lightrag.llm.ollama import ollama_model_complete, ollama_embed
 from lightrag.utils import EmbeddingFunc
 
 import nest_asyncio
