+++ lightrag/storage.py
@@ -6,6 +6,8 @@ from dataclasses import dataclass
 from typing import Any, Union, cast, Dict
 import networkx as nx
 import numpy as np
+import pipmaster as pm
+
 from nano_vectordb import NanoVectorDB
 import time
