+++ lightrag/types.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from typing import List
+
+class GPTKeywordExtractionFormat(BaseModel):
+    high_level_keywords: List[str]
+    low_level_keywords: List[str]
