+++ requirements.txt
@@ -1,22 +1,20 @@
 accelerate
-aioboto3
 aiofiles
 aiohttp
-aioredis
+redis
 asyncpg
 configparser
 
 # database packages
 graspologic
 gremlinpython
-hnswlib
 nano-vectordb
 neo4j
 networkx
 
+# TODO : Remove specific databases and move the installation to their corresponding files
+# Use pipmaster for install if needed
 numpy
-ollama
-openai
 oracledb
 pipmaster
 psycopg-pool
@@ -33,14 +31,12 @@ python-dotenv
 python-pptx
 pyvis
 setuptools
-# lmdeploy[all]
 sqlalchemy
 tenacity
 
 
 # LLM packages
 tiktoken
-torch
 tqdm
-transformers
 xxhash
+
